CMS-673: Refactor and use regular season save endpoint for winter dates
